refactor: remove date picker demo module
