+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow-integration-tests/src/main/java/com/vaadin/flow/component/datepicker/entity/Person.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.datepicker.demo.entity;
+package com.vaadin.flow.component.datepicker.entity;
 
 import java.time.LocalDate;
