+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow-integration-tests/src/test/java/com/vaadin/flow/component/datepicker/DatePickerIT.java
@@ -25,21 +25,27 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.component.datepicker.testbench.DatePickerElement;
-import com.vaadin.tests.ComponentDemoTest;
+import com.vaadin.flow.testutil.TestPath;
+import com.vaadin.tests.AbstractComponentIT;
 import com.vaadin.testbench.TestBenchElement;
+import com.vaadin.testbench.TestBenchTestCase;
 
 import static org.junit.Assert.assertTrue;
 
 /**
  * Integration tests for the {@link DatePickerViewDemoPage}.
  */
-public class DatePickerIT extends ComponentDemoTest {
+@TestPath("vaadin-date-picker-test-demo")
+public class DatePickerIT extends AbstractComponentIT {
 
     private static final String DATEPICKER_OVERLAY = "vaadin-date-picker-overlay";
+    private TestBenchTestCase layout;
 
     @Before
     public void init() {
+        open();
         waitForElementPresent(By.tagName("vaadin-date-picker"));
+        layout = this;
     }
 
     @Test
@@ -310,10 +316,5 @@ public class DatePickerIT extends ComponentDemoTest {
 
         $("button").id("Locale-US").click();
         Assert.assertEquals("3/8/20", localePicker.getInputValue());
-    }
-
-    @Override
-    protected String getTestPath() {
-        return ("/vaadin-date-picker-test-demo");
     }
 }
